requirements.txt
azure-storage-blob==1.4.0
requests==2.21.0

